fix(migrations): disable migration for dev

+++ packages/backend/src/_migrations/1698876391788-optimize-migration.ts
@@ -1,16 +1,28 @@
 import { MigrationInterface, QueryRunner } from "typeorm";
+import { appLogger } from "../util";
+import { domifaConfig } from "../config";
 
-export class AutoMigration1698876391788 implements MigrationInterface {
-  name = "AutoMigration1698876391788";
+export class OptimizeInteractionsMigration1698876391788
+  implements MigrationInterface
+{
+  name = "OptimizeInteractionsMigration1698876391788";
 
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(
-      `DROP INDEX "public"."IDX_3bc72392a39f586374f0f7d577"`
-    );
+    if (domifaConfig().envId === "dev") {
+      appLogger.info("[MIGRATION] Skipped");
+      return;
+    }
+
+    appLogger.info("[MIGRATION] Delete deleted interactions");
+    await queryRunner.query(`DELETE from interactions where event = 'delete'`);
+
+    appLogger.info("[MIGRATION] DROP columns previousValue & event");
     await queryRunner.query(
       `ALTER TABLE "interactions" DROP COLUMN "previousValue"`
     );
     await queryRunner.query(`ALTER TABLE "interactions" DROP COLUMN "event"`);
+
+    appLogger.info("[MIGRATION] Create new indexes");
     await queryRunner.query(
       `CREATE INDEX "idx_structure_statut" ON "usager" ("structureId", "decision") `
     );
@@ -19,6 +31,15 @@ export class AutoMigration1698876391788 implements MigrationInterface {
     );
     await queryRunner.query(
       `CREATE INDEX "idx_interactions" ON "interactions" ("structureId", "usagerUUID", "dateInteraction") `
+    );
+
+    appLogger.info("[MIGRATION] DROP useless indexes");
+    await queryRunner.query(
+      `DROP INDEX "public"."IDX_3bc72392a39f586374f0f7d577"`
+    );
+
+    await queryRunner.query(
+      `ALTER TABLE "interactions" DROP CONSTRAINT "FK_495b59d0dd15e43b262f2da8907"`
     );
   }
